@@ -2,21 +2,34 @@
 {
     using Grace.AspNetCore.Hosting;
     using Microsoft.AspNetCore.Hosting;
+    using Microsoft.Extensions.Hosting;
+    using Microsoft.Extensions.Logging;
+    using Singularity.Microsoft.DependencyInjection;
 
     using System.IO;
-
 
     public class App
     {
         public static void Main()
         {
-            new WebHostBuilder().UseGrace()
-                                .UseKestrel()
-                                .UseIISIntegration()
-                                .UseContentRoot(Directory.GetCurrentDirectory())
-                                .UseStartup<Startup>()
-                                .Build()
-                                .Run();
+            new HostBuilder().UseGrace()
+                             .ConfigureWebHostDefaults(whb =>
+                             {
+                                 whb.UseKestrel()
+                                    .UseIIS()
+                                    .UseStartup<Startup>();
+                             })
+                             //.UseServiceProviderFactory(new SingularityServiceProviderFactory())
+                             //.ConfigureLogging((ctx, logging) =>
+                             //{
+                             //    if (ctx.HostingEnvironment.IsDevelopment())
+                             //    {
+                             //        logging.AddDebug();
+                             //    }
+                             //})
+                             .UseContentRoot(Directory.GetCurrentDirectory())
+                             .Build()
+                             .Run();
         }
     }
 }
